fix 7715 permissions types
