+++ packages/permissions-kernel-snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/snap-7715-permissions.git"
   },
   "source": {
-    "shasum": "gsWbJM9WyU3yXbs4GfTcCNintvE+nGf36RJuFnoLVEg=",
+    "shasum": "ocSnSMt7++eqCQESyv2uUpPsnH8QolZi57AH+9pCCT4=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
