+++ packages/permissions-provider-snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/snap-7715-permissions.git"
   },
   "source": {
-    "shasum": "cs5Z3PZ+3PSCyVB5k3Kd/d3+asuJGqgxrzccSmIi7A0=",
+    "shasum": "K2W5eEelITcAbQC1/6+NhVV6cSIfRf6V4UaloWgD+PA=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
