+++ packages/shared/src/types/index.ts
@@ -1,4 +1,6 @@
-export * from './7715-permissions';
+export * from './7715-permissions-request';
+export * from './7715-permissions-response';
+export * from './7715-permissions-types';
 export * from './snap-permission-registry';
 export * from './error';
 export * from './snap';
